edit button added

+++ server/public/scripts/client.js
@@ -5,6 +5,7 @@ function onReady(){
     refreshToDoList();
     $('#enter-btn').on('click', addToDo);
     $('.to-do-list').on('click', '.delete-btn', deleteTask);
+    $('.to-do-list').on('click', '.edit-btn', editTask);
 };
 
 
@@ -12,16 +13,18 @@ function deleteTask(){
     console.log('in deleteTask()');
     taskID = $(this).data('id');
     console.log('ID for delete: ', taskID);
-    $.ajax({
-		method: 'DELETE',
-		url: `/toDo/${taskID}`
-	})
-		.then(function () {
-			refreshToDoList();
-		})
-		.catch(function (error) {
-			alert(`Delete Function Error: ${error}`);
-		});
+    if(confirm('Are you sure you want to delete the task??')){
+        $.ajax({
+            method: 'DELETE',
+            url: `/toDo/${taskID}`
+        })
+        .then(function () {
+            refreshToDoList();
+        })
+        .catch(function (error) {
+            alert(`Delete Function Error: ${error}`);
+        });
+    }
 }
 
 function addToDo(){
@@ -98,7 +101,7 @@ function renderToDoList(toDoList) {
                     class="time"
                     value="${toDo.time}"
                     readonly>
-                <button class="edit" data-index="${i}" data-id="${toDo.id}">Edit</button>
+                <button class="edit-btn" data-index="${i}" data-id="${toDo.id}">Edit</button>
                 <button class="delete-btn" data-id="${toDo.id}">Delete</button>
             </div> 
     `);
